refactor: クラスを統合
