+++ 07.ls_object/ls_command.rb
@@ -6,23 +6,35 @@ require_relative './print_file'
 require_relative './file_list'
 
 class LsCommand
-  def initialize(command)
-    @command = command
+  def initialize(argv)
+    @options = parse_options(argv)
+    @file_path = argv.first || './'
   end
 
   def run
-    file_list = FileList.new(@command.file_path, all: @command.options[:all])
-    file_list.reverse if @command.options[:reverse]
+    file_list = FileList.new(@file_path, all: @options[:all])
+    file_list.reverse if @options[:reverse]
     print_file = PrintFile.new(file_list)
-    return print_file.print_file_path if @command.contain_file_name?
+    return print_file.print_file_path if File.file?(@file_path)
 
-    if @command.options[:long]
+    if @options[:long]
       print_file.print_long_files
     else
       print_file.print_files
     end
   end
+
+  private
+
+  def parse_options(argv)
+    options = { all: false, reverse: false, long: false }
+    opt = OptionParser.new
+    opt.on('-a') { |v| options[:all] = v }
+    opt.on('-r') { |v| options[:reverse] = v }
+    opt.on('-l') { |v| options[:long] = v }
+    opt.parse!(argv)
+    options
+  end
 end
 
-command = Command.new(ARGV)
-LsCommand.new(command).run
+LsCommand.new(ARGV).run
